+++ src/by/it/group310902/kasperets/lesson06/A_LIS.java
@@ -1,0 +1,43 @@
+package by.it.group310902.kasperets.lesson06;
+
+import java.io.FileNotFoundException;
+import java.io.InputStream;
+import java.util.Scanner;
+
+public class A_LIS {
+
+    public static void main(String[] args) throws FileNotFoundException {
+        InputStream stream = A_LIS.class.getResourceAsStream("dataA.txt");
+        A_LIS instance = new A_LIS();
+        int result = instance.getSeqSize(stream);
+        System.out.print(result);
+    }
+
+    int getSeqSize(InputStream stream) throws FileNotFoundException {
+        // подготовка к чтению данных
+        Scanner scanner = new Scanner(stream);
+        // читаем длину последовательности
+        int n = scanner.nextInt();
+        // читаем всю последовательность
+        int[] sequence = new int[n];
+        for (int i = 0; i < n; i++) {
+            sequence[i] = scanner.nextInt();
+        }
+
+        // инициализируем массив для динамического программирования
+        int[] dp = new int[n];
+        int maxLength = 0;
+
+        for (int i = 0; i < n; i++) {
+            dp[i] = 1;
+            for (int j = 0; j < i; j++) {
+                if (sequence[i] > sequence[j]) {
+                    dp[i] = Math.max(dp[i], dp[j] + 1);
+                }
+            }
+            maxLength = Math.max(maxLength, dp[i]);
+        }
+
+        return maxLength;
+    }
+}
