+++ src/by/it/group310902/kasperets/lesson06/B_LongDivComSubSeq.java
@@ -1,0 +1,43 @@
+package by.it.group310902.kasperets.lesson06;
+
+import java.io.FileNotFoundException;
+import java.io.InputStream;
+import java.util.Scanner;
+
+public class B_LongDivComSubSeq {
+
+    public static void main(String[] args) throws FileNotFoundException {
+        InputStream stream = B_LongDivComSubSeq.class.getResourceAsStream("dataB.txt");
+        B_LongDivComSubSeq instance = new B_LongDivComSubSeq();
+        int result = instance.getDivSeqSize(stream);
+        System.out.print(result);
+    }
+
+    int getDivSeqSize(InputStream stream) throws FileNotFoundException {
+        // подготовка к чтению данных
+        Scanner scanner = new Scanner(stream);
+        // читаем длину последовательности
+        int n = scanner.nextInt();
+        // читаем всю последовательность
+        int[] sequence = new int[n];
+        for (int i = 0; i < n; i++) {
+            sequence[i] = scanner.nextInt();
+        }
+
+        // инициализируем массив для динамического программирования
+        int[] dp = new int[n];
+        int maxLength = 0;
+
+        for (int i = 0; i < n; i++) {
+            dp[i] = 1;
+            for (int j = 0; j < i; j++) {
+                if (sequence[i] % sequence[j] == 0) {
+                    dp[i] = Math.max(dp[i], dp[j] + 1);
+                }
+            }
+            maxLength = Math.max(maxLength, dp[i]);
+        }
+
+        return maxLength;
+    }
+}
