+++ src/by/it/group310902/kasperets/lesson06/Lesson06Test.java
@@ -1,4 +1,4 @@
-package by.it.a_khmelev.lesson06;
+package by.it.group310902.kasperets.lesson06;
 
 import org.junit.Test;
