+++ src/by/it/group310902/kasperets/lesson06/dataA.txt
